Renamed SettingsSMSVerify to SettingsVerifySecretCode so we can use it for verifying both SMS and email codes. Moved VoterEmailAddressEntry and VoterPhoneVerificationEntry into components/Settings folder.
